# backendAtNode/openapi.yaml
openapi: 3.0.3
info:
  title: Agriprix Log Management API
  description: |
    A comprehensive log management system API that allows organizations to upload, process, and query various log formats.
    
    ## Features
    - Multi-format log parsing (JSON Lines, Syslog, Nginx Error, CLF)
    - Organization-based log isolation
    - Role-based access control
    - File upload and batch processing
    - Advanced search and filtering
    - Manual log creation (Manager only)
    - Flattened metadata storage for optimized querying
    - Environment-based CORS configuration
    
    ## Authentication
    All endpoints except `/auth/login`, `/auth/signup`, and `/health` require a valid JWT token in the Authorization header.
    
    ## CORS Configuration
    The API supports configurable CORS origins via the `ALLOWED_ORIGINS` environment variable.
    Supports multiple origins separated by commas (e.g., `http://localhost:5173,http://frontend:5173`).
    
    ## Docker Deployment
    The API supports both development and production Docker deployments:
    - Development: `docker compose --profile dev up --build`
    - Production: `docker compose --profile prod up --build`
    
  version: 1.0.0
  contact:
    name: Agriprix Tech Team
    email: tech@agriprix.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.agriprix.com
    description: Production server
paths:
...
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user and create or link to an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
                - organization
                - role
              properties:
                email:
                  type: string
                  format: email
                  example: john.doe@company.com
                username:
                  type: string
                  minLength: 3
                  example: johndoe
                password:
                  type: string
                  minLength: 6
                  example: securePassword123
                organization:
                  type: string
                  example: Acme Corporation
                role:
                  $ref: '#/components/schemas/Role'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  organization:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 507f1f77bcf86cd799439011
                      name:
                        type: string
                        example: Acme Corporation
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with username, email, and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  format: email
                  example: john.doe@company.com
                password:
                  type: string
                  example: securePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Invalid credentials
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/signup endpoints
  schemas:
    Role:
      type: string
      enum: [viewer, manager]
      description: User role determining access permissions
      example: viewer
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: Validation failed
    ValidationError:
      type: object
      properties:
        detail:
          type: string
          example: Validation errors
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
                example: Email is required
              param:
                type: string
                example: email
              location:
                type: string
                example: body
  responses:
    ValidationError:
      description: Validation error in request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            detail: Validation errors
            errors:
              - msg: Valid email is required
                param: email
                location: body
              - msg: Password is required
                param: password
                location: body
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Internal server error
...